Moved all functions to one file.
Shifting focus to the GUI rather than writing to YAML. Writing to YAML needs finishing, but isn't the focus. Can be completed after the GUI. As long as variables can save values from the GUI, writing them to YAML will be a matter of using the optimal library/ugly-writing the values in.